+++ App/Il2CppOutputProject/Source/il2cppOutput/Microsoft.MixedReality.Toolkit3.cpp
@@ -3632,7 +3632,7 @@ struct Response_tCFDFF51E587CC0A680707D3ADF4A21893D4DB82A_marshaled_pinvoke
 {
 	int32_t ___U3CSuccessfulU3Ek__BackingField_0;
 	char* ___U3CResponseBodyU3Ek__BackingField_1;
-	Il2CppSafeArray/*I1*/* ___U3CResponseDataU3Ek__BackingField_2;
+	Il2CppSafeArray/*NONE*/* ___U3CResponseDataU3Ek__BackingField_2;
 	int64_t ___U3CResponseCodeU3Ek__BackingField_3;
 };
 // Native definition for COM marshalling of Microsoft.MixedReality.Toolkit.Utilities.Response
@@ -3640,7 +3640,7 @@ struct Response_tCFDFF51E587CC0A680707D3ADF4A21893D4DB82A_marshaled_com
 {
 	int32_t ___U3CSuccessfulU3Ek__BackingField_0;
 	Il2CppChar* ___U3CResponseBodyU3Ek__BackingField_1;
-	Il2CppSafeArray/*I1*/* ___U3CResponseDataU3Ek__BackingField_2;
+	Il2CppSafeArray/*NONE*/* ___U3CResponseDataU3Ek__BackingField_2;
 	int64_t ___U3CResponseCodeU3Ek__BackingField_3;
 };
 
@@ -6469,7 +6469,7 @@ struct Exception_t_marshaled_pinvoke
 	char* ____source_12;
 	SafeSerializationManager_t4A754D86B0F784B18CBC36C073BA564BED109770 * ____safeSerializationManager_13;
 	StackTraceU5BU5D_t855F09649EA34DEE7C1B6F088E0538E3CCC3F196* ___captured_traces_14;
-	Il2CppSafeArray/*INT*/* ___native_trace_ips_15;
+	Il2CppSafeArray/*NONE*/* ___native_trace_ips_15;
 };
 // Native definition for COM marshalling of System.Exception
 struct Exception_t_marshaled_com
@@ -6488,7 +6488,7 @@ struct Exception_t_marshaled_com
 	Il2CppChar* ____source_12;
 	SafeSerializationManager_t4A754D86B0F784B18CBC36C073BA564BED109770 * ____safeSerializationManager_13;
 	StackTraceU5BU5D_t855F09649EA34DEE7C1B6F088E0538E3CCC3F196* ___captured_traces_14;
-	Il2CppSafeArray/*INT*/* ___native_trace_ips_15;
+	Il2CppSafeArray/*NONE*/* ___native_trace_ips_15;
 };
 
 // System.Reflection.BindingFlags
@@ -18941,7 +18941,7 @@ IL2CPP_EXTERN_C IL2CPP_METHOD_ATTR float MathUtilities_DegreesToRadians_mBB46C07
 	{
 		// return (float)(degrees * Mathf.Deg2Rad);
 		double L_0 = ___degrees0;
-		return (((float)((float)((double)il2cpp_codegen_multiply((double)L_0, (double)(0.01745329238474369))))));
+		return (((float)((float)(float)((double)il2cpp_codegen_multiply((double)L_0, (double)(0.01745329238474369))))));
 	}
 }
 // System.Single Microsoft.MixedReality.Toolkit.Utilities.MathUtilities::RadiansToDegrees(System.Single)
@@ -29277,7 +29277,7 @@ IL_0077:
 	{
 		// int index = (int)(x * 0.5f);
 		int32_t L_21 = V_5;
-		V_6 = (((int32_t)((int32_t)((float)il2cpp_codegen_multiply((float)(((float)((float)L_21))), (float)(0.5f))))));
+		V_6 = (((int32_t)((int32_t)(int32_t)((float)il2cpp_codegen_multiply((float)(((float)((float)L_21))), (float)(0.5f))))));
 		// Vector3 forward = forwardList[index];
 		List_1_tFCCBEDAA56D8F7598520FB136A9F8D713033D6B5 * L_22 = ___forwardList4;
 		int32_t L_23 = V_6;
